Chapter_08_RAG/Basic_RAG_EXPLAIN/requirements.txt
chromadb>=1.5
pypdf>=6.0
flask>=3.0
requests>=2.30
python-dotenv>=1.0
